+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `log-level` flag and default to `error` level.
+
 ### Changed
 
 - Use apiextensions-application to remove CAPI dependency.